Fixed no-scaling mode
My implementation sometimes ignored the no-scaling setting, because some
window managers ignore requests for window sizes larger that the screen
size. After this commit, in no-scaling mode, images are always displayed
at 100% if no-scaling is enabled, regarless of the window size. I also
disabled in-window-centering in this mode, because this led to funny
displayements at least in xfwm.

Related to issue #4