Full support for the old configuration file format
I felt that forcing users to migrate their configuration files wasn't a
good idea, so I implemented a parser for the configuration which inverts
the meaning of command line flags when they also appear in the
configuration. So a `-f -n 5' configuration file will, if pqiv is called
with `pqiv -f foo', result in pqiv NOT starting in fullscreen again.

Using a flag two times will still NOT inverse the meaning of the second
use. So `-ff' still is fullscreen, contrary to pqiv ≤ 1.0