Better handling of images not loaded yet
Big images need some time to load. If a user navigates to a big image
before it has been loaded, the program froze until it was. This is
mostly still the case, but now the user sees an information message on
what is happening and can still toggle fullscreen and/or quit.

fixes #6