Renamed `.qiv-select' directory to `.pqiv-select'
This is a change Debian's maintainer added from the earliest package on.
Back then I did not want to adopt it, because I wanted pqiv to feel as
much as qiv as possible. Now, with qiv back in the game and pqiv
consistently naming itself pqiv in all other places, I feel renaming the
directory and thereby removing the last occurrence of `qiv' on its own
is the right thing to do.