Store a scaled version of each image to make redraws faster
This is important for fading, see bug #15. I decided to use another
image surface for this, because during fades, there are 2 images which
must both be drawn: The old one AND the new one. Now
last_visible_image_surface stores the old image, and the new variable
stores the scaled copy of the new one.