Use weak references when passing around BOSNodes
This is required because a thread/callback might be invoked after an
image has already been removed from the tree. The implementation is very
simple: bostree got two new functions, bostree_node_weak_ref() and
unref, respectively. They increment and decrement a counter, as a normal
reference counter would do. The references are weak in the sense that if
a node is removed, it becomes invalid and the weak reference will
resolve to NULL.