Remove file_sorted_tree
Now that the nested-symlink problem is solved via a separate tree, the
only improvement brought by the sorted list is to help users who
unintentionally specified a parameter twice or who symlinked an image
multiple times. But both cases admit intentional use, and we prevented
users from that before. Therefore I think removing this feature is a
good idea.