Do not exit if the first image given to -r fails to load
The image loader initialization fails if none of the images successfully
loads. Until now, it then printed an error message & quit. With this
change, it just silently fails. Any error messages are outputted only
when it is clear that all images from the command line and -r have been
processed.