Exit with success if the help was displayed
Related to pull request #22.

Other programs exit with status 0 if invoked without parameters / when
the help is shown. pqiv still returns 1 if all images failed to load,
which is the case a scripts couldn't test for on its own, so I think
that's okay to do the same and return 0.