Added --browse
See bug #25. qiv had this option for some time now, and with --max-depth
this is actually useful if pqiv is used as a handler for images in a
file manager, where users expect to be able to switch between images
from the same directory.

(Note: This does work with --sort and --shuffle. The first parameter is
 still the first image displayed. ☺)