Workaround for feathered edges
cairo feathers image borders (i.e. draws the border with a lower alpha
value). It's not a drawing problem; if the scaled image surface is saved
to a PNG file, that file also has the feathered border.

That is not a serious problem on its own, but it becomes one if the
checkerboard pattern is drawn: Especially with dark images, the light
border around the image resulting from it is clearly visible and
irritating. On the other hand, if the checkerboard pattern is only drawn
1px into the image, transparent images look as if they had a black
border in windowed mode.

As a workaround, I draw the checkerboard pattern as before in windowed
mode, but 1px into the image in fullscreen.

A more permanent solution would be nice.

What I've found out so far:

 * The problem lies in cairo: If drawing a transparent eps using
   libspectre, it can also be observed
 * It is likely related to different round-offs while scaling
 * Though not related to different rounding in rasterized/vector
   images: If I draw an unscaled checkerboard image to an image-surface
   and then scale that, the problem can still be observed.