Add explicit X11 threads initialization
See https://bugs.freedesktop.org/show_bug.cgi?id=82630

It would certainly be better to avoid X11 calls from separate threads
alltogether, and to my best knowledge, I do that in the main
application. But something in the backends seem to use gtk_* calls as
well, and I haven't even been able to tell where and which one so far.
(I suspect cairo.) Fixing this in another way and reverting this is a
TODO.