Guess the file type of images that doesn't match any extension
Previously, the default (gdk-pixbuf) was used. The same goes for memory
images. What we do is to use GIO's content type system to autodetect the
type based on the file's content. From the docs:

> A content type is a platform specific string that defines the type of a
> file. On UNIX it is a mime type like "text/plain" or "image/png". On
> Win32 it is an extension string like ".doc", ".txt" or a perceived
> string like "audio". Such strings can be looked up in the registry at
> HKEY_CLASSES_ROOT.

We then use a MIME-type converter from GIO to convert that into a common
format and detect based on that. gdk-pixbuf comes with a function to add
the supported types to a GtkFileFilter (we actually already used that);
for the other file types, the corresponding MIME types were manually
added.