Revert to using the display name for sorting, not absolute file name
When sort mode is enabled, pqiv used to remove files from the tree if
they were deleted in --watch-directories mode. This feature was
introduced in b112304 to close bug #7. In ed754a1 I found that a problem
with this behaviour was that we index the tree by *display name*, while
the inotify callback naturally only has access to the *absolute path
name*, so the mechanism could only react to files that were specified by
absolute path names and that were handled by a loader that doesn't alter
the display name (i.e. not poppler/spectre, which are multi-page
capable). In ed754a1, this problem was solved by altering the index to
use the absolute path name instead.

A problem with that, that only revealed itself later, is, that sorting
of multi-page files (foo.pdf[1] .. foo.pdf[n]) doesn't work anymore -
the pages are now reversed. Instead of doing a hack (like reversing the
insertion of pages if sorting is enabled) I felt that reverting the
change would be the right thing. This also means that removing files
from the tree as they vanish from the disk is no longer an option, as it
is buggy.

I think that with the changes introduced with ed754a1 this is no longer
problematic: Files that were already loaded are now removed if they
are deleted. And, of course, files that were not cannot be loaded
anyways. So the only harm that comes back with this commit is that the
displayed image count is not correct sometimes.