Preliminary implementation of a toggleable shuffle mode
See bug #30 for an alternative, lengthy explanation. To make shuffle
mode toggleable, we can no longer simply choose random keys for the file
tree, but must instead shuffle in the relative_image_movement()
function.

For users, it is important that shuffle mode doesn't display images that
they have already seen. Also, it is desired that forward and backward
image movement works predictable despite the images being shuffled.

The proposed solution uses a linear congruential PRNG whose parameters
are chosen such that each sequence of random numbers forms a full cycle,
i.e. that no number is chosen twice until all other numbers have been
generated. We invent the formula to allow backwards movement. If an
image is viewed, it is marked as "seen". If the user leaves it via
backward movement, this mark is again removed. If a user moves forward
through the images and shuffle mode is enabled, only images that don't
have the mark set are considered. Once all images have the mark set, it
is removed from all images and the process starts over again.

I am not sure whether this works out, and also not sure if removing the
"seen" flag is a good idea.