Implementation of a toggleable shuffle mode
See bug #30 for the backgrounds of this change. This commit introduces a
toggleable shuffle mode, where a user can switch off shuffle mode to
view some images in sequential order and then switch it on again and
continue at where he initially disabled it. Commit 0dc33dd contains an
alternative implementation, that also counts the sequentially visited
images as "seen" and is based on a completely different implementation.
Especially it doesn't require us to maintain a separate list of images,
but instead uses an invertible PRNG. See bug #30 for why I eventually
rejected that approach.

This variant works as follows: pqiv maintains a list of images that have
been seen in random order. Relative image movements in shuffle mode take
place within this list. If a movement crosses either end of the list,
the list is dynamically filled up with randomly chosen images that are
not yet present in the list. (In a list deciding this is O(n)!) To speed
things up a little, we do not choose random numbers until we find one
that is not yet present, but instead choose a random one and then
sequentially walk through the image tree until we find one that isn't.
This also generates eqidistributed permutations. (I didn't prove that,
but large samples look as if it does.) Once this list is full, it is
flushed and regenerated. The list contains weak references to the
images, such that the mechanisms that insert/remove files as they are
altered on the disk still works.