Fixes for lazy loading with the wand backend
The magick wand backend uses a global mutex since it isn't reentrant. It
might call functions that again call functions in the backend though,
when allocating space for a new file: pqiv would try to load the file if
it is the first one, resulting in a dead-lock. A similar problem occurs
if the user quits the program while the backend is still processing a
file. This commit fixes both bugs.