Make inotify work on multi-page images
This fixes a bug where PDF files would not be properly reloaded when
they changed on disk. The issue was that we use a filebuffer structure
to access the files (because old versions of Poppler did not accept
arbitrary input-streams, and spectre/wand do not do at all) that is not
per file_t, but per physical file, as identified by the stucture's
->file_name. inotify triggers once per loaded image, issues a reload,
but since the file is still in cache when the reload of a page happens -
because other pages are still loaded - the file is never reread from
disk. The fix is to first unload every file with the same ->file_name in
the image loader thread and only then reload the image in question.

This introduces a ->force_reload flag on file_t's, that will activate
the unloading logic described above. A side effect of this commit is
that unload_image() is now properly called from the loader thread and
not from the inotify callback, which runs in the main thread.