Added --sort-key=mtime to sort files by modification time instead of by name
This introduces a new file_t attribute, ->sort_name, that contains the
variant of the file name that the file is sorted by. The option works by
prepending the filename with its mtime. While this might seem like
overkill for now, this solution allows for easy extension with further
sort keys.