Make features configurable at compile time
This is an initial attempt to make certain features optional, like it
used to be in pqiv 1. Stripping features that one doesn't use from pqiv
makes the binary considerably smaller.

I'll still have to figure out which features are best for this (this
commit contains the external commands - which are large - and shuffling,
which doesn't make much of a difference (~3-4 kiB with debug symbols))
and whether it would be better to refactor the optional features into
their own files to make the separation simpler/better testable/more
obvious.