Fix shuffle mode skipping one image per cycle
This was a logical issue: The shuffle mode implementation worked by
building a list of images in relative_image_pointer(), traversing this
list to navigate between images in shuffle mode, and emptying the list
once it was complete, i.e. contained all the images. The problem here is
that due to preloading, the last image of a cycle would have been
determined upon loading of the first (since the previous image is held
in memory), resulting in it never being viewed.

The solution is to furthermore maintain a bit vector of flags indicating
whether an image has actually been viewed in shuffle mode. Once all
images have, the shuffle list is emptied. If relative_image_pointer()
finds that all images have been used but not all have been viewed, it
wraps at the ends of the list.

I use two counter variables to avoid having to search the list too
often: One for its length, one for the number of images that have been
viewed. (A linear search is still required to search the current image
in the list, I couldn't come up with a reasonably fast solution for that
yet..)