Disable GTK's transparent scaling on HiDpi monitors
The approach is quite simple: Scale all sizes and positions retrieved
from GTK by GDK_SCALE, and scale all size and position requests and draw
functions by 1/GDK_SCALE.

The intent of GDK_SCALE is to improve viewability on High DPI screens.
This is a useful feature for the jump dialog and the yellow info box.
Hence, do not scale the jump dialog and scale up the font size of the
yellow info box.

See also bug #75.