Do not use RPATH anymore, instead search for libraries within pqiv
This should resolve a bug that surfaced at least on Gentoo: If a system
defaults to ld's --enable-new-dtags feature, then -rpath writes to the
RUNPATH tag instead of to the RPATH tag. dlopen() ignores RUNPATH, so
the libraries are no longer found if they aren't in the expected
directories.

One solution would have been to enforce --disable-new-dtags, but since
this will most likely sooner or later be deprecated, I thought I'd go
through with this slightly less portable approach.