Be a bit less humble about pqiv
This is related to #79. pqiv's self-description led to people believing
that pqiv was almost on par with qiv's feature set and that qiv was
superior. Of course, pqiv has *more* features than qiv. The updated
description should make clearer what the intentions and use cases of
pqiv are. I also added remarks on its idea of caching and preloading to
make clear that it does not aim for a particularly low memory footprint.

Also, towards the same end, the libav backend is compiled in by default
if it is available.