Replace image surface with native surface for scaled image, fix fading
This commit replaces the temporary_scaled_image_surface with a native
surface, massively increasing performance for translations of
transformed images in --low-memory mode. This is a follow-up change to
issue #99.

While at it, I've also fixed the combination of fading and slideshows if
the fade interval is longer than the slideshow interval, as far as this
is possible. See #74.