Enable cursor-auto hide by default
This was initially not possible because motion events are expensive in
X11 at least. This commit mitigates the issue by unsetting the motion
event bit after receiving a single motion event. This way, only the
initial event that notifies us that the user moved the cursor is sent.
After each second, pqiv checks if the cursor was moved. As soon as it
stops, it hides the cursor and reenables the motion event request.

Cursor auto-hide is enabled in windowed and montage mode, but disabled
in normal fullscreen, by default.