Queue actions and continue chains with an idle handler
This allows to use

    add_file(foo)
    goto_file_byname(foo)

from stdin to go to file foo in a more reliable fashion than before. It
still won't be completely reliable though: The add_file() action will
still be called asynchronously, and goto_file_byname will be called as
soon as the main thread is back to idle, so there still is a race - but
that's intentional, since actions shall not block for a long time.

(In case you need full reliability, use set_status_output() and check
whether the second call succeeds.)

This change also improves reliability of action chains.

Handling the "when to continue an action chain" better than currently is
still a TODO however.