Replace static build CI with Github actions
This disables the self-hosted CI that created static builds for windows
and linux in favor of Github's actions.

Static builds in linux have been broken for some time now. The reason
for that is that GTK 3.x does not want to support static linking, and
while it has been possible to hack static builds together by maintaining
a few patches, the maintenance effort is becoming overwhelming.

Static builds for windows work flawlessly, thanks to the awesome mxe.cc
project, but I am not aware of anyone actually using them. If needed,
they can be easily brought back using Github actions and a custom Docker
container.