Fix wrap around of shuffled image list
If the shuffled image list doesn't include all images yet, pqiv attempts
to extend it to fulfil a relative image movement request. The logic in
the code checked once whether the list can be extended at all, and then
unconditionally attempt to extend it by a sufficient number of images.
That's flawed, since e.g. a user might request to move by 5 images when
there are only 3 images left that have never been viewed. The correct
logic would be to add those 3, and then wrap around to the second image.
This commit applies that fix.

This fixes #176.