allow keybinding of special keys with shift modifier
During configuration of the keybindings, removing the shift modifier
while uppercasing the key did not work for special keys like <Delete>
that can't be uppercased.  So only uppercase keys that actually have a
proper uppercase variant.

Now a combination like '<Shift><Delete>' can be read as a keybinding,
but it later breaks in the input handling, which always removes the
shift modifier on the received keyval before any further processing.

So instead of always removing the shift modifier during input
handling, call gdk_keymap_translate_keyboard_state() to figure out
which modifiers already have been consumed while getting the
keyval.  (eg. when the keyval is 'A', the shift modifier has already
been consumed and can be removed; when the keyval is '<Delete>', a
possible shift modifier should be passed through for further
processing)

The fallback code "if there is no keybinding for 'A', look for a
binding for '<Shift>a' instead" remains unchanged.