Use Pango to render infobox text.
Cairo's built-in Unicode handling is limited and does not properly
render certain scripts. Pango has close integration with Cairo and
provides complete Unicode character rendering.

The --font argument is also added to allow the desired font to be
passed to Pango from the command line or configuration. 'Sans-Serif'
is used by default with the preexisting size of 12. In the case that the
desired size is too large, it will be scaled down as before.