Do not use GDK Wayland module
pqiv has some rendering issues with this module, the same it had 8 years
ago when I first experimented with it. Since the x11 backend still works
just fine with Wayland, I see no need to fix this right now.

The issues are quite visible when running with GDK_BACKEND=wayland. See
also the TODO at the end of the draw callback.