Point fix: Align memory for libav sws buffer
The libav backend had a bug in software scaling. It set alignment
requirements for a buffer used for software scaling to 16 bytes, which
worked fine in past versions, but results in crashes nowadays with
videos that do not happen to naturally align as needed. This commit sets
the alignment to 64 byes, under the (unverified) assumption that ffmpeg
now uses vector instructions with that alignment requirement.

This change fixes the crash observed in #247 for the example file.